Discard old beta model working with probabilities
