+++ neasqc_wp61/6_Beta.sh
@@ -19,5 +19,4 @@ echo "K values for KNN algorithm: ${k[*]}";
 echo "output: $output";
 
 echo "running beta_1"
-
-python3.10 ./data/data_processing/use_beta_1_bert.py -tr ${train} -te ${test} -k ${k[*]} -o ${output}
+python3.10 ./data/data_processing/use_beta_1.py -tr ${train} -te ${test} -k ${k[*]} -o ${output}
