+++ neasqc_wp61/6_Beta_Neighbours.sh
@@ -20,4 +20,4 @@ echo "output: $output";
 
 echo "running beta_1"
 
-python3.10 ./data/data_processing/use_beta_1_bert.py -tr ${train} -te ${test} -k ${k[*]} -o ${output}
+python3.10 ./data/data_processing/use_beta_1.py -tr ${train} -te ${test} -k ${k[*]} -o ${output}
